requirements.txt
django-allauth>=0.44.0
djangorestframework>=3.14.0,<4.0
drf-yasg>=1.21.0,<2.0
